feat(#31): add active chapter logic
