+++ src/components/Chapters.tsx
@@ -1,8 +1,17 @@
 import React from 'react';
 import { P } from './P';
 import { useChapterCreate, useChaptersByCoverId } from '../apis/useChapters';
+import { Chapter } from '../types/chapters';
 
-export default function Chapters({ coverId }: { coverId: number }) {
+export default function Chapters({
+  coverId,
+  activeChapter,
+  setActiveChapter,
+}: {
+  coverId: number;
+  activeChapter: Chapter | undefined;
+  setActiveChapter: React.Dispatch<React.SetStateAction<Chapter>>;
+}) {
   const { data: chapters } = useChaptersByCoverId(coverId);
   const { mutate: createChapter } = useChapterCreate(coverId);
   return (
@@ -20,8 +29,19 @@ export default function Chapters({ coverId }: { coverId: number }) {
           }}
         />
       </div>
-      {chapters?.map((c) => {
-        return <div key={c.id}>{c.title}</div>;
+      {chapters?.map((chapter) => {
+        return (
+          <P
+            size="base"
+            className={
+              'mt-4 cursor-pointer ' + (chapter.id === activeChapter?.id ? '' : 'text-secondary-2')
+            }
+            key={chapter.id}
+            onClick={() => setActiveChapter(chapter)}
+          >
+            {chapter.title}
+          </P>
+        );
       })}
     </>
   );
