+++ src/types/api.ts
@@ -1,0 +1,3 @@
+export type Response<T> = {
+  data: T;
+};
