+++ src/types/chapters.ts
@@ -1,0 +1,11 @@
+export type Chapter = {
+  id: number;
+  title: string;
+  public: boolean;
+  blockIds: string[];
+};
+
+export type CreateChapter = {
+  coverId: number;
+  title: string;
+};
